requirements.txt
Django==3.1
psycopg2-binary==2.8.6
Pillow==7.1
django-filter==2.4.0
python-decouple==3.4
